+++ MyRuns1/src/main/AndroidManifest.xml
@@ -25,6 +25,7 @@
         <activity android:name=".ListviewActivity"></activity>
         <activity android:name=".GpsActivity"
             android:label="Map"></activity>
+        <activity android:name=".DisplayEntryActivity"/>
     </application>
 
 </manifest>
